requirements.txt
dj-database-url==0.4.1
Django==1.9.8
djangorestframework==3.5.3
gunicorn==19.6.0
httplib2==0.9.2
oauth2client==4.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
whitenoise==3.2
